Moviddo código relacionado ao inicio da fase para outro arquivo (StartGameController) e ocultado botões de jogo na tela inicial
