+++ Assets/StartGameController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 35f93e58ffeb36d4890a34ad2436fed9
+timeCreated: 1452352622
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
